Add files via upload
In the final project, the task of predicting customer churn from the "ниединогоразрыва.ком" network was set. The goal of the project is to build a machine learning model to solve the binary classification problem, where the first group of people will include people who want to break the contract with the store chain, and the second group -- customers who plan to continue using the company's services. For future work, three machine learning models will be used to solve the problem.