Add files via upload
The service for the sale of used cars "Not beaten, not painted" is developing an application to attract new customers. In it, you can quickly find out the market value of your car. Historical data is at your disposal: technical characteristics, equipment and prices of cars. Build a model to determine the value of your car.