Add files via upload
The presented project presents analytics on datasets that were provided as assignment to the prefabricated project. Two models from sklearn library were used in the study, namely random forest and logistic regression. Also, the data were processed and outliers and outliers were removed. It is worth noting that during the study it turned out that the enrichment efficiency calculations were not done correctly.