Add files via upload
As part of the project, I researched customer churn from a bank. A dataframe was provided for the study, which provides various customer data. The data takes integer values such as scoring and age. Also some data take fractional values like salary, account balance. And categorical variables are encountered, such as target attribute -- exited and numofproducts. So from the information provided by the marketers, we are to investigate the situation that is prevailing in the bank by creating and training machine learning models. Various libraries will be used for the task at hand, as well as the MO models themselves in order to maximize the accuracy of the result obtained.