Add files via upload
In this project, I will analyze a dataset that consists of several variables that take the form of strings such as name, platform, genre, rating, integer values such as year_of_release, critic_score, and fractional values: _sales, user_score. Also in the rating column you can find values that are responsible for age restrictions on games, for example E10+ means that the game is available for people over 10 years old, and EU rating means "Early childhood" and this product is available for everyone.